final/requirements.txt
# AgriTech ML Project Dependencies

# Core ML Libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
xgboost>=1.6.0
joblib>=1.2.0
prophet>=1.1.0

# Web Framework
flask>=2.2.0
flask-cors>=4.0.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Frontend
streamlit>=1.20.0

# HTTP Requests & APIs
requests>=2.28.0
aiohttp>=3.8.0

# Translation
indictrans>=1.0.0
transformers>=4.21.0

# Environment Variables
python-dotenv>=0.19.0

# Additional Utilities
calendar
datetime
geopy>=2.3.0  # For location coordinates
